rename mpi to massPayment and mpiCollection to massPaymentCollection, move MassPayment to Collection
